Release 2.0.0

+++ app/src/main/java/com/tyganeutronics/myratecalculator/database/Currency.kt
@@ -1,6 +1,8 @@
 package com.tyganeutronics.myratecalculator.database
 
-abstract class Currency constructor(var id: String, var rate: Double) {
+import java.math.BigDecimal
+
+abstract class Currency constructor(var id: String, var rate: BigDecimal) {
 
     abstract fun getSign(): String
     abstract fun getName(): Int
